+++ src/image_loading.rs
@@ -2,7 +2,7 @@ use crate::color::Color;
 use crate::image::Image;
 use crate::image_loading::ImageWrapperError::{GraphicsLibError, ImageLibError};
 use crate::GraphicsError;
-use image::{DynamicImage, ImageError, ImageFormat};
+use image_lib::{DynamicImage, ImageError, ImageFormat};
 use std::io::{BufRead, Seek};
 use std::path::Path;
 use thiserror::Error;
@@ -21,12 +21,12 @@ pub fn load_image<R: BufRead + Seek>(
     r: R,
     format: ImageFormat,
 ) -> Result<Image, ImageWrapperError> {
-    convert_image(image::load(r, format).map_err(|e| ImageLibError(e))?)
+    convert_image(image_lib::load(r, format).map_err(|e| ImageLibError(e))?)
 }
 
 /// Load an image (png, bmp, etc) as [Image]
 pub fn open_image<P: AsRef<Path>>(path: P) -> Result<Image, ImageWrapperError> {
-    convert_image(image::open(path).map_err(|e| ImageLibError(e))?)
+    convert_image(image_lib::open(path).map_err(|e| ImageLibError(e))?)
 }
 
 ///Convert a [DynamicImage] from the `image` crate into an [Image]